Refactor file structure by organizing source and header files
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/BrandonTang89/HanabiTelegramBot?shareId=XXXX-XXXX-XXXX-XXXX).
